Move upload loaded files logic into the uploadContext
