+++ frontend/contexts/UploadContext.js
@@ -1,4 +1,4 @@
-import React, { createContext, useContext, useState, useEffect } from 'react'
+import React, { createContext, useContext, useEffect, useState } from 'react'
 import { useRouter } from 'next/router'
 
 import _ from 'lodash'
@@ -14,7 +14,7 @@ export const UploadContext = createContext({
   allFiles: [],
   loadedFiles: [],
   loadFiles: () => {},
-  uploadFile: async (repo, path, content, csrf) => true,
+  uploadLoadedFiles: () => {},
   setPersistenceScope: () => {},
 })
 
@@ -88,11 +88,31 @@ export default function UploadContextProvider(props) {
     }
   }
 
-  const uploadFile = uploadFileToGitea
+  const uploadLoadedFiles = async repo => {
+    // there's a race condition happens on gitea when making several request to the upload endpoint
+    // a hacky/awful solution to get around it is simulating a scheduler with setTimeout
+    const delay = 1000
+    const res = await Promise.all(
+      loadedFiles.map(async (file, idx) => {
+        const content = sessionStorage.getItem(`loadedFile_${file.name}`)
+        setTimeout(async () => {
+          return await uploadFileToGitea(repo, file.name, content, csrf)
+        }, delay * idx)
+      }),
+    )
+
+    console.log(res)
+  }
 
   return (
     <UploadContext.Provider
-      value={{ loadedFiles, loadFiles, uploadFile, allFiles, setPersistenceScope }}
+      value={{
+        loadedFiles,
+        loadFiles,
+        uploadLoadedFiles,
+        allFiles,
+        setPersistenceScope,
+      }}
     >
       {props.children}
     </UploadContext.Provider>
